lab 4

+++ lab_03/lab_03_02_00/main.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
-#include "helper.h"
+#include "matrix_io.h"
 
 int main(void)
 {
     int a[LEN*2][LEN];
-    int n, m;
+    size_t n, m;
 
     printf("Введите количество строк и столбцов в матрице\n");
-
-    if (scanf("%d%d", &n, &m) != 2 || n <= 0 || m <= 0)
+    
+    if (scanf("%zu%zu", &n, &m) != 2 || n <= 0 || m <= 0)
     {
         printf("Ошибка ввода размера матрицы\n");
         return ERROR_INPUT;
